+++ src/options/components/single-processor-tab.tsx
@@ -223,7 +223,14 @@ export function SingleProcessorTab() {
                 generateWebIconKit: formatOptions.ico.generateWebIconKit
               }
             : undefined,
-        png: targetFormat === "png" ? { tinyMode: formatOptions.png.tinyMode } : undefined
+        png:
+          targetFormat === "png"
+            ? {
+                tinyMode: formatOptions.png.tinyMode,
+                cleanTransparentPixels: formatOptions.png.cleanTransparentPixels,
+                autoGrayscale: formatOptions.png.autoGrayscale
+              }
+            : undefined
       },
       resize: { mode: "none" }
     }